+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 * Minor changes have been made to text that is output to the console.
 * The fuzz effect of spectres, and the player’s weapon when they have a partial invisibility power-up, are now paused when freeze mode is on.
+* The player’s weapon will now lower slightly when they drop down from a greater height.
 
 ---